Provide a node.js module that can be installed locally and required
